Warning cleanup
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23711 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/Streamline/avtStreamlineRenderer.h
@@ -95,7 +95,7 @@ class avtStreamlineRenderer : public avtCustomRenderer
     // lighting props
     virtual void GlobalLightingOn();
     virtual void GlobalLightingOff();
-    virtual void GlobalSetAmbientCoefficient(const float);
+    virtual void GlobalSetAmbientCoefficient(const double);
     virtual void SetSpecularProperties(bool,double,double,
                                        const ColorAttribute&);
